+++ dvs-daemon.py
@@ -36,41 +36,8 @@ from collections import deque
 from Queue import Queue
 
 import infoplus_dvs
+import dvs_util
 
-def setup_logging(default_path='logging.yaml',
-    default_level=logging.INFO, env_key='LOG_CFG'):
-    """
-    Setup logging configuration
-    """
-
-    path = default_path
-    value = os.getenv(env_key, None)
-    if value:
-        path = value
-    if os.path.exists(path):
-        with open(path, 'rt') as config_file:
-            log_config = yaml.load(config_file.read())
-        logging.config.dictConfig(log_config)
-    else:
-        logging.basicConfig(level=default_level)
-
-def load_config(config_file_path='config/dvs-server.yaml'):
-    """
-    Setup logging configuration
-    """
-
-    global config
-
-    if os.path.exists(config_file_path):
-        try:
-            with open(config_file_path, 'rt') as config_file:
-                config = yaml.load(config_file.read())
-        except Exception as e:
-            print "Fout in configuratiebestand. Foutmelding: %s" % e
-            sys.exit(1)
-    else:
-        print "Configuratiebestand '%s' niet aanwezig" % config_file_path
-        sys.exit(1)
 
 def main():
     """
@@ -102,17 +69,10 @@ def main():
     args = parser.parse_args()
 
     # Laad configuratie:
-    load_config(args.configFile)
+    config = dvs_util.load_config(args.configFile)
 
     # Stel logging in:
-    log_config_file = None
-
-    # Check of er een logger configuratie is opgegeven:
-    if 'logging' in config and 'log_config' in config['logging']:
-        log_config_file = config['logging']['log_config']
-    
-    # Stel logger in adhv config:
-    setup_logging(log_config_file)
+    dvs_util.setup_logging(config)
 
     # Geef logger instance:
     logger = logging.getLogger(__name__)
@@ -678,7 +638,7 @@ class GarbageThread(threading.Thread):
 # Injector thread:
 class InjectorThread(threading.Thread):
     """
-    Thread die verantwoordelijk is voor garbage collection
+    Thread die verantwoordelijk is voor verwerken van DVS injecties
     """
 
     logger = None
@@ -701,20 +661,14 @@ class InjectorThread(threading.Thread):
         while True:
             try:
                 # Ontvang injection dict
-                trein_dict = client_socket.recv_pyobj()
+                trein_dict = client_socket.recv_json()
                 
                 self.logger.debug("Nieuwe injectie: %s", trein_dict)
                 counters['injecties'] += 1
 
-                # Stuur response naar injector
-                client_socket.send_pyobj(True)
-
                 # Converteer ontvangen dict naar 
                 trein = infoplus_dvs.parse_trein_dict(trein_dict, True)
-
-                # Bepaal rit ID. Prefix 'i' om overlap met InfoPlus
-                # DVS ID's te voorkomen.
-                rit_id = 'i%s' % trein.rit_id
+                rit_id = trein.rit_id
 
                 # Voeg trein toe aan stores:
                 with locks['trein']:
@@ -730,8 +684,12 @@ class InjectorThread(threading.Thread):
                 station_store[trein.rit_station.code][rit_id] = trein
                 trein_store[rit_id][trein.rit_station.code] = trein
 
-            except Exception:
-                self.logger.exception("Fout tijdens verwerken injectie")
+                # Stuur response naar injector
+                client_socket.send_json({'result': True})
+
+            except Exception as e:
+                self.logger.exception("Fout tijdens verwerken injectie: %s", e)
+                client_socket.send_json({'result': False, 'error': str(e)})
 
 if __name__ == "__main__":
     main()